feat: 제안 연봉 필드 타입 변경

+++ src/main/java/com/oaosis/samak/domain/analysis/dto/response/AnalysisItemDetailResponse.java
@@ -26,7 +26,7 @@ public record AnalysisItemDetailResponse(
         String companyName,
 
         @Schema(description = "제안 연봉", example = "3000000")
-        Long salary,
+        BigDecimal salary,
 
         @Schema(description = "생성 날짜")
         LocalDateTime createdAt
